mesa: don't validate pipeline in GL debug contexts
This is causing problems with glretrace and various traces I've seen,
including F1 2015 and PCSX2. I haven't carefully debugged what's going
on, but the games do work fine when tracing, which most likely points to
the validation logic not being suitable for desktop GL. I've left in the
ability to flip it in with a debug flag.

Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>
Cc: "12.0" <mesa-stable@lists.freedesktop.org>
Cc: Ian Romanick <idr@freedesktop.org>