nvc0: do not follow the push path for indexed draws
This fixes a lot of flickering in The Talos Principle on Fermi+
(tested on both GF119 and GK106). My guess is that some indices
and/or some vertex attributes are incorrectly set when emitting
data directly into the command stream rather than uploading a
user VBO.

The NV50 driver also doesn't use the push path for indexed draws
which makes confident enough with this fix. Furthermore, this
doesn't introduce any regressions with piglit.

As a side-effect, this also improves performance with The Talos
Principle (+~20% fps) and probably with other games which use indexed
draws. This is because uploading a user VBO is more efficient than
pushing data directly into the pushbuf.

There is still an issue in this game (sometimes walls become green),
but this seems to be totally unrelated to the first issue (disabling
the push path doesn't help here).

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=90513
Signed-off-by: Samuel Pitoiset <samuel.pitoiset@gmail.com>
CC: "11.1 11.2" <mesa-stable@lists.freedesktop.org>